ref: Move all consumer retry code into the library
This change moves the BasicRetryPolicy code from Snuba into the
shared library. This policy is likely to be useful for other users
of this library.
